requirements.txt
scipy==1.0.1
matplotlib==2.2.2
scikit_image==0.13.1
numpy==1.14.2
tensorflow==1.6.0
opencv_python==3.4.0.12
Pillow==5.1.0
